financial-document-analyzer-debug/requirements.txt
click==8.1.8
crewai==0.130.0
crewai-tools==0.47.1
fastapi==0.110.3
uvicorn==0.29.0
pydantic==2.12.5
openai==2.24.0
langchain-community==0.3.31
python-dotenv==1.2.1
python-multipart==0.0.22
numpy==1.26.4
pandas==2.2.2
pillow==10.3.0